Fix Quasar SFPU compile errors: use SFPU_CALL with template ITERATIONS.
Quasar SFPU functors (`_calculate_X_` / `calculate_X`) take `ITERATIONS`
only as a template parameter (zero runtime args). Several call sites
were passing it as a runtime arg through `SFPU_CALL_FN` / `SFPU_CALL_MODE`,
which does not compile on Quasar. Switch them to `SFPU_CALL` /
`SFPU_CALL_MODE` with `ITERATIONS` in the TEMPLATES tuple.

Also:
- Restore the dropped `is_fp32_dest_acc_en == false` static_assert in the
  Quasar rsqrt wrapper.
- Drop the `static_cast<int>(vector_mode)` mismatch in the Quasar ternary
  macro; the params function takes `VectorMode` directly.
- Remove the now-unused `num_sfpu_iterations` runtime variable / dispatcher
  parameter from the Quasar test files (abs, square, square_trisc3, rsqrt,
  nonlinear).

Addresses Copilot review comments on PR #46329.

+++ tt_metal/hw/ckernels/quasar/metal/llk_api/llk_sfpu/llk_math_eltwise_ternary_sfpu_macros.h
@@ -52,7 +52,7 @@ inline __attribute__((always_inline)) void _sfpu_ternary_check_and_call_(
         dst_index_in1,
         dst_index_in2,
         dst_index_out,
-        static_cast<int>(vector_mode),
+        vector_mode,
         std::forward<Args>(args)...);
 }
